Add msstore productId to export manifest and to wingetutil interop manifest object (#4934)
This change was needed as part of the effort to make exported manifest
(from winget download msstore apps) to be re-parsed and uploaded to
winget rest source.
 
Updated the some of the manifest schemas to add "required" for some
required fields previously missing.

Also, during my cross validation with the manifest schemas, I found
ExpectedReturnCode::Custom was added in manifest but not actually
updated in code. So I updated the ExpectedReturnCode map in code with
this change too.