Don't put ARP version data into tracking catalog (#4964)
## Change
A previous change was made to block overlapping ARP ranges being put
into the index. The fact that this was potentially impactful to the
tracking catalog was put off until now.

This change removes ARP version information (when present) from the
manifest before placing it into the tracking catalog. We are not using
this data, and I don't see a reason that we would. Attempting to
preserve it would probably require removing all existing ARP version
information when adding/updating the tracking manifest.

## Validation
Added a test that to insert an overlapping range into the tracking
catalog.