Add missing top-level exception handling for arg validation (#5111)
The CLI arg validation code has gotten more complicated and now there is
more room for exceptions to occur. For example, if there is an error
when reading the admin settings or group policies that determine if an
argument is available.

This adds generic exception handling at the top-level of arg validation
so that we can exit gracefully and show an error, instead of just
crashing.

Partial fix for #5098. This addresses the no output, but not the cause
of the error

###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/winget-cli/pull/5111)