Add configuration environment (#5139)
## Change
This change adds a `ConfigurationEnvironment` type which is a property
of a `ConfigurationUnit`. It defines the environment in which the unit
should be run. This currently encompasses two properties; the security
context and the processor.

The values are parsed from the metadata, but the object is considered
authoritative at runtime and when serializing. They are also inherited
in schema 0.3, allowing the entire set to be defined in a single
environment in it's metadata.

The code that was using the `securityContext` metadata has been updated
to use the environment property instead. It also uses the unique
environment calculation function for the set to more efficiently
determine which contexts are present.