Add support to export all installed packages in winget configure export (#5156)
The feature is in parity with `winget export` command. Except the
exported file is in winget configuration yaml format. The exported file
is ready to be used with winget configure commands.

New additions:
- When exporting a specific package (existing functionality), in
addition we'll search the source to find the package before export. If
the package is not from well known source, the source is exported as
well. Supports exporting the version like `winget export` too.
- Added `--all` to export all installed packages. We reuse the workfow
in `winget export` to collect packages to export. For non well known
sources, the sources are exported as well. `--all` cannot be used with
other specific package export arguments.

Manually validated and added e2e tests.