Update Newtonsoft.Json to be same version across all projects (#5158)
According to the NuGet management tools in Visual Studio,
`Newtonsoft.Json` had different versions being used across different
projects within the solution. This PR bumps them all to the same
version, moving some from `13.0.2` up to `13.0.3`. In the cases where
the version specification was added, it is due to the package being a
transient dependency