Update YamlDotNet to Latest (#5191)
I know that NuGet packages are typically not updated unless there are
CVEs, important bug fixes, or other specific need to. However, when
reviewing the list of NuGet packages installed, I saw that YamlDotNet
was several versions out of date. This package directly affects the
parsing of manifests within WinGet and the associated utilities. Given
that, I thought it best to update it, especially given the fact that
1.10 is preparing to go through the release cycle, which would give
these changes enough time to be tested in the 1.11 previews and testing
cycles.

This change only directly updates the top-level project -
`WinGetUtilInterop`. I'm expecting that all the other projects which
have this NuGet package as a transient dependency will pick up this new
version, as all of them have a Project Dependency on
`WinGetUtilInterop`.

The projects which consume this change -

### Top-Level
* WinGetUtilInterop
### Transitive
* AppInstallerCLIE2ETests
* Tool\IndexCreationTool
* Tool\LocalhostWebServer
* WinGetSourceCreator
* WinGetUtilInterop.UnitTests