Validate URLs map to unique hashes (#5233)
<!-- To check a checkbox place an "x" between the brackets. e.g: [x] -->
- [x] I have signed the [Contributor License
Agreement](https://cla.opensource.microsoft.com/microsoft/winget-pkgs).
- [x] This pull request is related to an issue.
  - Resolves #2852

## Description:
This PR adds validation for URL to SHA256 mappings. It is presumed that
a single URL should only return a single SHA256, so if a manifest has
different SHA256 for the same URL, this will now throw a validation
*error*. This will also help reduce potential load on the Validation
Pipelines at winget-pkgs as the `InstallerSha256` mismatch will be
caught during manifest validation and the validation will exit before it
proceeds to installer validation.

Additionally, as noted in the initial issue, it would be unlikely (but
possible) for two different URLS to point to the same SHA256. Validation
was added for this case, but will only throw a validation *warning*. It
is also necessary for this to be a warning so that this change does not
become a breaking change.


![{D367385B-6718-46F4-818D-4198F75D36C9}](https://github.com/user-attachments/assets/a5cfdce4-16c3-4c0e-8db0-b543fee3123a)

-----

###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/winget-cli/pull/5233)

---------

Co-authored-by: Muhammad Danish <mdanishkhdev@gmail.com>
Co-authored-by: JohnMcPMS <johnmcp@microsoft.com>