Fix logging channel setting (#5261)
## Issue
The user setting to control the logging channels was no longer being
respected after a previous change to enable all channels during
initialization (and to send the initialization logs to the debug
stream).

## Change
`EnableChannel` for logging has always been semantically "add the given
channel(s) to the list of enabled channels".

Every place that we were calling `EnableChannel` was really semantically
wanting to set the list of enabled channels rather than add to it.

This change creates a new function `SetEnabledChannels` that behaves
that way; it simply sets the enabled channel mask to the given value.
All callers of `EnableChannel` were changed to use this new function.