Remove support for arm32 (#5290)
This removes support for building winget for arm32. All references to
the platform in the build configurations are removed, but I left some in
the code - like the list of applicable architectures for arm32. This
does not touch anything related to the support of packages for arm32,
only building winget itself.

#5289

###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/winget-cli/pull/5290)