Add validation option for portable installer type in archives (#5237)
## Description
For the `portable` installer type, the `winget-pkgs` repository has a
policy of not allowing scripts as installers. This includes `.bat`,
`.ps1`, and `.cmd` files.

In addition to the policy at `winget-pkgs`, the client itself is not
designed to handle portable files that are not `.exe` types. Issues have
been filed asking for WinGet to support these in the future, but it
currently does not -
* https://github.com/microsoft/winget-cli/issues/5083
* https://github.com/microsoft/winget-cli/issues/5086
* https://github.com/microsoft/winget-cli/issues/5087

This PR adds a validation option which validates the `RelativeFilePath`
within a zip file is always a `.exe`, but can be extended to include
additional formats as the requests above are addressed. This was added
as a validation option (similar to the Schema Header Validation) as
private repositories may want to not have the file type restrictions
that `winget-pkgs` does - as described in these issues -
* https://github.com/microsoft/winget-cli/issues/5084
* https://github.com/microsoft/winget-cli/issues/5085

This should allow for the validation pipeline at `winget-pkgs` to
automatically reject manifests that contain `NestedInstallerFiles` that
are not supported, while allowing maintainers of private repositories to
suppress the restriction by turning the validation option to false.