DSC v3 Export (#5319)
## Change
Adds a more generic `GetAllUnits(Async)` export function to the
`ConfigurationProcessor`. This produces full configuration units, which
may then differ from the given unit's type. This models the DSC v3
`Exporter` resource type. The implementation of this function prefers
the full implementation from the processor, but if not provided it can
fall back to the previous `GetAllSettings` function and synthesize
results.

Implements both the "all settings" and "all units" functionality for the
DSC v3 processor. This in turn is done through the only DSC v3 option,
`export`.