Only download during COM download (#5348)
Cherry pick #5327 to 1.10

## Issue
When `download` was added, the COM download and install split for
dependencies was accidentally implemented. But it was validated only
with download in mind, so the download part worked for download intent.
With dependencies, the split basically attempted to install the
dependencies twice, causing a failure on the second attempt due to
internal assertions.

## Change
For the COM download command, explicitly force that we are only
downloading.