Fixes for pipeline stability (#5359)
## Issue
Due to an issue with dsc.exe creating child processes (which after much
investigation appears to be a race that only occurs on older OS
versions), we get arbitrary failures.

We also occasionally get spurious failures of configuration tests.

## Change
Move to `windows-2025` to avoid the process creation failures from
dsc.exe.
Move to Windows 11 SDK 26100 as 22000 is not installed by default on
`windows-2025`.
Flush output at the end of a configuration unit info dump.
Add missing interface to manifest (@yao-msft was right).
Use `pwsh` task at least once to ensure that it is present.