DSC v3 adapter support (#5302)
## Change
If we don't find a resource in DSC v3 at the top level, look in all of
the adapters for it. `dsc.exe` does this automatically when using the
resource, so we only need to do it when finding the resource details.

Also improves the diagnostics handling and fixes an issue with the type
of `implementedAs`.

## Validation
Adds an E2E test that uses the test v2 resource through the v3
processor.