Lower Schema Header Validation to include 1.7+ manifests (#5330)
<!-- To check a checkbox place an "x" between the brackets. e.g: [x] -->
- [x] I have signed the [Contributor License
Agreement](https://cla.opensource.microsoft.com/microsoft/winget-pkgs).
- [ ] This pull request is related to an issue.

Many contributors to the community repository may not upgrade to schema
1.10 unless necessary to do so, as there were no additional fields in
the manifest that are usable in the community repository. In addition to
manifest creation tools lagging behind a little bit (as of this PR, both
Komac and WingetCreate are still at version 1.9 manifests), some
submitters use mechanisms outside of these tools which use older
manifest versions.

This first PR changes the schema header validation so that it happens
for any manifest schema 1.7 or newer to cover a larger range of
submissions.

I have manually validated this change against *every* manifest currently
in the community repository. Three had to be corrected, and now none are
failing validation.