Remove most checked-in dependency libraries (#5390)
This removes the code for most of the dependency libraries that was
checked in, in favor of getting them from vcpkg. This required enabling
vcpkg for some projects, like Fuzzing.

The dependencies removed are:
* `catch2`
* `JsonCppLib`
* `Valijson`
* `yamlcpp`
* `detours` (transitive dependency of `UndockedRegFreeWinRT`)

From `PureLib`, also removed the transitive dependency `dotzlib`. We
don't need a replacement for this one.

Moving `catch2` to vcpkg also brought in a major update (v2->v3), which
required some changes in the included headers and namespaces.

From `cpprestsdk`, removed some of the code we don't need, like samples
and build pipelines. We should remove all of `cpprestsdk`, but it is not
as straightforward because we modified it for certificate pinning, so we
need to apply a patch on top of the original code from the port.

`PureLib` and `sfs-client` weren't removed because there is no
equivalent in vcpkg.


###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/winget-cli/pull/5390)