Remove clones of transitive dependencies available from vcpkg (#5399)
Follow-up to #5390

We have copies of `PureLib` and `UndockedRegFreeWinRT` in the repo that
themselves include code from their dependencies (`catch2`, `zlib`). This
PR removes that code in favor of using vcpkg to fetch the dependencies.

For `PureLib`, after removing `zlib`, there are only headers left. So I
replaced the `.vcxproj` with a `.vcxitems`

For `UndockedRegFreeWinRT`, I removed all the test code along with the
copy of catch2.

Updated the READMEs for the subtrees, although I'm not certain a subtree
update would still work...
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/winget-cli/pull/5399)