Ensure all vcpkgs use our custom triplets (#5423)
After the change to have our dependencies use vcpkgs instead of checking
in a clone of the code, we got BinSkim alerts due to some of them not
using our custom triplets with the compilation flags that we need. This
PR ensures that all projects include the `vcpkg.props` that makes CMake
use our custom triplets instead of the default `{arch}-windows` ones.

It is unclear to me why this happened. For example, why the test
installer exe was building the vcpkgs at all, but the logs show that as
one of the root projects for the issue.

Besides importing `vcpkg.props` in more projects, this also does a
little refactor of the triplets to reduce duplication and have the
required flags in a single place
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/winget-cli/pull/5423)