Update sample projects to 1.10.340 (#5427)
This updates the sample projects for consuming our nuget packages to use
v1.10.340.

For the winget out-of-proc package, we can now remove the manual
references from the project files because the package now has a targets
file.

For the configuration in-proc package, I'm not updating the version
because we stopped publishing it.


###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/winget-cli/pull/5427)