Create a vcpkg overlay for cpprestsdk and libyaml (#5424)
This PR adds a local port overlay so that we can have our own custom
vcpkg ports. In particular we use it to:
1. Apply a patch to cpprestsdk for certificate pinning, without cloning
the whole project
2. Apply a patch to libyaml for a security vulnerability, as there is no
updated version with the fix

These ports are mostly a copy of the official ports, but with the
patches we want applied. Unfortunately the original port recipes were
already applying some patches, and we had to include them here because
vcpkg provides no mechanism to apply a single additional patch on top of
an existing port.

Notes for the future:
* We could potentially use the port overlay to create ports of the other
libraries we have clones of, without having all their code checked in
* For libyaml, instead of applying a patch we could simply change the
base commit to one that already has that change
* For libyaml, we could also try to contribute the patch to the official
port
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/winget-cli/pull/5424)