Place DSCv3 resource manifests in the package (#5433)
## Change
Adds a method to write all manifests to a directory with one command.
Uses that to write all DSCv3 resource manifests and add them to the
package files.

Also changes the "module" name for the dev package to add ".Dev" and
updated export to use the shared definition of the module name. This
involved a minor refactor to remove static globals and unnecessary
string constructions.

Finally, moves to using the DSC v3 `UserSettingsFile` resource rather
than the v2 version.