Attempt to use DSCv3 resources even if they were not found (#5443)
## Change
Even if not found, attempt to use DSCv3 resources. This is largely to
avoid PowerShell/DSC#786, which only seems to affect dsc.exe's ability
to run processes when listing adapted resources.

Also applies the `DiagnosticTraceEnabled` state to the elevated
processor. This is set when the config logging area is enabled and set
to verbose.