Use more recent version for libyaml (#5455)
We have a Component Governance alert for libyaml. There is no release of
libyaml with this issue fixed, so the guidance was to apply the patch
manually, and that's what I did when moving to vcpkg, but that doesn't
play nicely with CG because it can't determine that the patch to fix the
vulnerability was applied.

Instead of manually patching, this PR uses a more recent commit of
libyaml (without an official release), which already has the changes we
want. It uses the same commit from the last time we did a subtree update
in #4583

---------

Co-authored-by: JohnMcPMS <johnmcp@microsoft.com>