Add ReleaseStatic configuration to WinGetServer project (#5477)
Each vcpkg port can only be installed with a single triplet at a time
(in a given install directory). Since we have different triplets for
static and non-static configurations, we need to ensure that all
projects use the same triplet for a given solution configuration.

This PR adds the ReleaseStatic configuration to the WinGetServer project
and sets the solution to use that when appropriate. It also includes a
bit of deduplication in the vcxproj.

Having WinGetServer use the triplet for the non-static configuration
when building for static, is causing errors in the internal build
because it deletes some of the installed triplets that other projects
later try to consume.
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/winget-cli/pull/5477)