Change how we determine which framework architectures to install (#5517)
Fixes #5020 

## Change
This change consolidates the choice of which framework architectures to
install to a single function and changes it from the current "whatever
`OSArchitecture` tells us + x86 and x64 if it was ARM64" to the set of
architectures that we see installed across all other frameworks. The
goal is to work around the fact that in some instances we are not
receiving an honest answer from `OSArchitecture`.

This does rely on the presence of at least one framework package of the
system architecture.

It also adds a "break in case of emergency" environment variable
`WINGET_PACKAGE_MANAGER_REPAIR_DEPENDENCY_ARCHITECTURES` that can be
used to control the set of architectures in the event of unexpected
issues.