Add text file with the release notes (#5519)
This adds a text file with the notes we add to each release here in
GitHub.
I intend to update the publishing script to include these notes
automatically so I don't have to manually copy them each time.

These notes are not intended to be a full changelog of all versions, but
rather just about the changes in the current version. (Although maybe we
should have a changelog for all major versions at least.)

It would be nice if we manage to keep the notes up to date every time we
make a change, but if we forget we can still edit the release after the
fact.
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/winget-cli/pull/5519)

---------

Co-authored-by: yao-msft <50888816+yao-msft@users.noreply.github.com>