Refresh PATH in configuration remoting server (#5511)
Added a hidden window to listen for UserPreferenceChanged event
(WM_SETTINGCHANGE) and update the process's PATH if needed.

Manually validated by running a configuration file with multiple
resources, change the env var during the run, and verified in logs file
that new path got included in runs after the change.