Support for DSC minimum version (#5525)
## Change
Adds a state machine for finding a version of DSC that meets our minimum
requirements. It first attempts to use the stable release, falling back
to the preview version if it is not new enough. This allows us to have
functional, if less than ideal, user experience while we wait for DSC to
ship a GA version.

Both callers (winget and PS module) are updated to handle the state
machine transition mechanics, which means they need to install
stable/preview versions when requested.

Also fixes some issues with the PS module install flow.