Repair Repair-WinGetPackageManager (#5568)
`Repair-WinGetPackageManager` is failing to install winget in some Win10
systems. The issue seems come from #5517 and be caused because in some
cases we fail to mark any one architecture as the right one to use for
dependencies. This PR is a partial revert of #5517 to add the previous
logic as a fallback.

May fix #5559 (but I haven't set up an environment to test it...)
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/winget-cli/pull/5568)