Improve COM server quiescing (#5652)
## Issue
When the package needs to be updated, the COM server does not currently
shut down in response to the signals that it receives. This is because
it only shuts down when all COM object references have been released. We
need to be more proactive in termination as the process will be
terminated shortly after we are notified either way.

## Change
This change refactors the existing shutdown signaling code out from
`ExecutionContext.cpp` into its own, public location. It then adds
another system for coordinating complex shutdown that is triggered by
the signaler.

Individual subsystems can register with the shutdown synchronizer with
functions for:
- Disable new work
- Cancel current work
- Wait for quiesce

and on shutdown initialization, all subsystems are called for each stage
in order. When the final subsystem returns from waiting, the existing
"all COM objects have been released" event is signaled and the process
will exit.

Additionally, the shutdown signal handler waits for shutdown
synchronization to complete in the window message function as after we
return from that we are expected to be gone shortly after. While this
will lead to some increased number of non-quiescing processes, we want
to do the best that we can to cooperatively exit.

> [!NOTE]
> Any actions being taken by callers outside of those described below
will **NOT** block the process from exiting.

### Package Management
Registers with synchronization on `PackageManager` creation. All of its
functions lead back to the `ContextOrchestrator` which has been
enlightened to handle the shutdown (and to operate as a contained set of
instances for test purposes).

- **Disable new work** causes the orchestrator to refuse to queue items
with an exception for new items and a cancellation for existing ones.
- **Cancel current work** causes all queued items to be cancelled, with
different handling based on current status.
- **Wait for quiesce** waits on an event for each queue that is signaled
when the queue is empty.

### Configuration
Registers with synchronization on `ConfigurationStaticFunctionsShim`
creation (the `IConfigurationStatics` for OOP use). All of its functions
are routed through `IConfigurationStaticsInternals` to a new shutdown
handler. It tracks all of the async token objects used in the
"background" sections of the `ConfigurationProcessor` functions. This
tracking is performed automatically by new types that manage that
tracking with the lifetime. It also enables non-async invocations to be
cancelled by providing a dummy promise into the existing system.

- **Disable new work** causes an error whenever a new async token object
is created.
- **Cancel current work** causes all tracked async tokens to have their
`Cancel` function invoked.
- **Wait for quiesce** waits until all async tokens are destroyed.