Update schema to 1.12 with Font InstallerType (#5687)
* Updates schema to 1.12 and adds type "font" as a valid InstallerType
and NestedInstallerType in the schema.
* Adds InstallerType Font to the manifest and associated Enum
conversions.
* Updated some of the NestedInstaller validation that had assumed
Portable type as the only type that could have multiple installer files
(fonts can also have many fonts in a package).
* Added validation for 1.12 fonts in the manifest and associated tests
for the schema version.

This PR does not add any functional support for Fonts, only the
validation and schema update so the new schema can be used and is valid
in the related WinGet tooling.

Tested:
* Passed the AppInstallerCLITests with [ManifestValidation] and ensured
a clean result.
* Added new tests for the v 1.12 manifest updated with the new installer
types in the test installer manifest to ensure Fonts are correctly
passing the validation.

Closes #5686 

<!-- To check a checkbox place an "x" between the brackets. e.g: [x] -->

- [x] I have signed the [Contributor License
Agreement](https://cla.opensource.microsoft.com/microsoft/winget-pkgs).
- [x] I have updated the [Release Notes](../doc/ReleaseNotes.md).
- [x] This pull request is related to an issue.

-----