Improve slow searches involving installed items (#5701)
Fixes #5629 
(and probably quite a few more list/upgrade some-query-provided
performance related issues)

While things could still be made better, the good case (where files were
already cached locally) of the previous behavior trying to complete
`winget list Micro` was taking 51 seconds. With this change, it takes 6
seconds on the debug build. Given that there is intentionally little to
no I/O and much of the CPU time is likely spent in our code, that should
speed up more with the release.

## Issue
The indexed `IPackage` implementation for v2 needed to acquire the
intermediate manifest to provide the response for `GetVersionKeys`.
Since this is used extensively to perform correlation between available
packages and local items during a search, any available packages found
by the query would then result in a download/disk read of the
intermediate manifest.

## Change
Introduces `IPackage::GetMultiProperty`, allowing code that just wanted
to get all of these values across all versions to do so more
efficiently. For the v2 index, this is how the data is already stored in
the database. For the other `IPackage` implementations, the data was
also present and is just aggregated from each version's output.