Add the ARP correlation entry to the context for portable installs (#5707)
Fixes #5688

## Issue
We were not recording ARP changes for portable installs, which prevented
the product code from being in the tracking database, which led to
inconclusive correlation results for packages with shared normalized
name+publisher.

## Change
Portable install flow extracts the `AppsAndFeaturesEntry` for the
install and puts it into the `CorrelatedAppsAndFeaturesEntries` context
item.